F_T30603 move MapBundle transformer to core
- adjust use statements
